Refactor provider outside of ncrack. This allows reuse across other bruteforcers.
