+++ google/detectors/credentials/ncrack/src/main/java/com/google/tsunami/plugins/detectors/credentials/ncrack/NcrackWeakCredentialDetector.java
@@ -33,10 +33,10 @@ import com.google.tsunami.plugin.PluginType;
 import com.google.tsunami.plugin.VulnDetector;
 import com.google.tsunami.plugin.annotations.PluginInfo;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.composer.WeakCredentialComposer;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.CredentialProvider;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.TestCredential;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.Top100Passwords;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.tester.CredentialTester;
+import com.google.tsunami.plugins.detectors.credentials.provider.CredentialProvider;
+import com.google.tsunami.plugins.detectors.credentials.provider.TestCredential;
+import com.google.tsunami.plugins.detectors.credentials.provider.Top100Passwords;
 import com.google.tsunami.proto.AdditionalDetail;
 import com.google.tsunami.proto.Credential;
 import com.google.tsunami.proto.DetectionReport;
