+++ google/detectors/credentials/ncrack/src/main/java/com/google/tsunami/plugins/detectors/credentials/ncrack/NcrackWeakCredentialDetectorBootstrapModule.java
@@ -20,9 +20,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.inject.Provides;
 import com.google.tsunami.plugin.PluginBootstrapModule;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.client.NcrackBinaryPath;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.CredentialProvider;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.Top100Passwords;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.tester.CredentialTester;
+import com.google.tsunami.plugins.detectors.credentials.provider.CredentialProvider;
+import com.google.tsunami.plugins.detectors.credentials.provider.Top100Passwords;
 import java.io.FileNotFoundException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
