+++ google/detectors/credentials/ncrack/src/main/java/com/google/tsunami/plugins/detectors/credentials/ncrack/composer/WeakCredentialComposer.java
@@ -19,9 +19,9 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Streams;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.CredentialProvider;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.TestCredential;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.tester.CredentialTester;
+import com.google.tsunami.plugins.detectors.credentials.provider.CredentialProvider;
+import com.google.tsunami.plugins.detectors.credentials.provider.TestCredential;
 import com.google.tsunami.proto.NetworkService;
 import java.util.Iterator;
 import java.util.List;
