+++ google/detectors/credentials/ncrack/src/main/java/com/google/tsunami/plugins/detectors/credentials/provider/Top100Passwords.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.tsunami.plugins.detectors.credentials.ncrack.provider;
+package com.google.tsunami.plugins.detectors.credentials.provider;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
