+++ google/detectors/credentials/ncrack/src/test/java/com/google/tsunami/plugins/detectors/credentials/ncrack/NcrackWeakCredentialDetectorTest.java
@@ -30,9 +30,9 @@ import com.google.tsunami.common.net.http.HttpClient;
 import com.google.tsunami.common.net.http.HttpClientModule;
 import com.google.tsunami.common.net.http.HttpStatus;
 import com.google.tsunami.common.time.testing.FakeUtcClock;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.CredentialProvider;
-import com.google.tsunami.plugins.detectors.credentials.ncrack.provider.TestCredential;
 import com.google.tsunami.plugins.detectors.credentials.ncrack.tester.CredentialTester;
+import com.google.tsunami.plugins.detectors.credentials.provider.CredentialProvider;
+import com.google.tsunami.plugins.detectors.credentials.provider.TestCredential;
 import com.google.tsunami.proto.AdditionalDetail;
 import com.google.tsunami.proto.Credential;
 import com.google.tsunami.proto.DetectionReport;
